refactoring folder handler and route api.auth
