--- api.event/controller/handler/handler.event.delete.go
@@ -1,8 +1,0 @@
-package handler
-
-import "github.com/gofiber/fiber/v2"
-
-// EventDelete handler
-func (s DbStruct) EventDelete(c *fiber.Ctx) error {
-	return nil
-}
