--- api.event/controller/handler/handler.event.get.all.go
@@ -1,8 +1,0 @@
-package handler
-
-import "github.com/gofiber/fiber/v2"
-
-// EventGetAll handler
-func (s DbStruct) EventGetAll(c *fiber.Ctx) error {
-	return nil
-}
