--- api.event/controller/handler/handler.event.get.one.go
@@ -1,8 +1,0 @@
-package handler
-
-import "github.com/gofiber/fiber/v2"
-
-// EventGetOne handler
-func (s DbStruct) EventGetOne(c *fiber.Ctx) error {
-	return nil
-}
