+++ api.event/controller/handler/handler.event.go
@@ -1,0 +1,28 @@
+package handler
+
+import "github.com/gofiber/fiber/v2"
+
+// EventDelete handler - v
+func (s DbStruct) EventDelete(c *fiber.Ctx) error {
+	return nil
+}
+
+// EventGetAll handler
+func (s DbStruct) EventGetAll(c *fiber.Ctx) error {
+	return nil
+}
+
+// EventGetOne handler
+func (s DbStruct) EventGetOne(c *fiber.Ctx) error {
+	return nil
+}
+
+// EventPost handler
+func (s DbStruct) EventPost(c *fiber.Ctx) error {
+	return nil
+}
+
+// EventUpdate handler
+func (s DbStruct) EventUpdate(c *fiber.Ctx) error {
+	return nil
+}
