--- api.event/controller/handler/handler.event.post.go
@@ -1,10 +1,0 @@
-package handler
-
-import (
-	"github.com/gofiber/fiber/v2"
-)
-
-// EventPost handler
-func (s DbStruct) EventPost(c *fiber.Ctx) error {
-	return nil
-}
