--- api.event/controller/handler/handler.event.update.go
@@ -1,8 +1,0 @@
-package handler
-
-import "github.com/gofiber/fiber/v2"
-
-// EventUpdate handler
-func (s DbStruct) EventUpdate(c *fiber.Ctx) error {
-	return nil
-}
