--- api.event/controller/handler/handler.evento.delete.go
@@ -1,1 +1,0 @@
-package handler
