--- api.event/controller/handler/handler.evento.get.all.go
@@ -1,1 +1,0 @@
-package handler
