--- api.event/controller/handler/handler.evento.get.one.go
@@ -1,1 +1,0 @@
-package handler
