--- api.event/controller/handler/handler.evento.post.go
@@ -1,8 +1,0 @@
-package handler
-
-import "github.com/gofiber/fiber/v2"
-
-// EventoGetOne handler responsável por tratar os dados e chamar o repo, para realizar o trás 1 registro de mesa.
-func (s DbStruct) EventoGetOne(c *fiber.Ctx) error {
-	return nil
-}
