--- api.event/controller/handler/handler.evento.update.go
@@ -1,1 +1,0 @@
-package handler
