--- api.event/controller/handler/handler.table.delete.go
@@ -1,8 +1,0 @@
-package handler
-
-import "github.com/gofiber/fiber/v2"
-
-// TableDelete handler responsável por tratar os dados e chamar o repo, para realizar o delete do post na tabela de mesas.
-func (s DbStruct) TableDelete(c *fiber.Ctx) error {
-	return nil
-}
