--- api.event/controller/handler/handler.table.get.all.go
@@ -1,8 +1,0 @@
-package handler
-
-import "github.com/gofiber/fiber/v2"
-
-// TableGetAll handler responsável por tratar os dados e chamar o repo, para realizar o get de todos os registros mesas.
-func (s DbStruct) TableGetAll(c *fiber.Ctx) error {
-	return nil
-}
