--- api.event/controller/handler/handler.table.get.one.go
@@ -1,8 +1,0 @@
-package handler
-
-import "github.com/gofiber/fiber/v2"
-
-// TableGetOne handler responsável por tratar os dados e chamar o repo, para realizar o trás 1 registro de mesa.
-func (s DbStruct) TableGetOne(c *fiber.Ctx) error {
-	return nil
-}
