+++ api.event/controller/handler/handler.table.go
@@ -1,0 +1,28 @@
+package handler
+
+import "github.com/gofiber/fiber/v2"
+
+// TableDelete handler responsável por tratar os dados e chamar o repo, para realizar o delete do post na tabela de mesas.
+func (s DbStruct) TableDelete(c *fiber.Ctx) error {
+	return nil
+}
+
+// TablePost handler responsável por tratar os dados e chamar o repo, para realizar as post na tabela de mesas.
+func (s DbStruct) TablePost(c *fiber.Ctx) error {
+	return nil
+}
+
+// TableUpdate handler responsável por tratar os dados e chamar o repo, para realizar updates no registro da mesa.
+func (s DbStruct) TableUpdate(c *fiber.Ctx) error {
+	return nil
+}
+
+// TableGetAll handler responsável por tratar os dados e chamar o repo, para realizar o get de todos os registros mesas.
+func (s DbStruct) TableGetAll(c *fiber.Ctx) error {
+	return nil
+}
+
+// TableGetOne handler responsável por tratar os dados e chamar o repo, para realizar o trás 1 registro de mesa.
+func (s DbStruct) TableGetOne(c *fiber.Ctx) error {
+	return nil
+}
