--- api.event/controller/handler/handler.table.post.go
@@ -1,10 +1,0 @@
-package handler
-
-import (
-	"github.com/gofiber/fiber/v2"
-)
-
-// TablePost handler responsável por tratar os dados e chamar o repo, para realizar as post na tabela de mesas.
-func (s DbStruct) TablePost(c *fiber.Ctx) error {
-	return nil
-}
