--- api.event/controller/handler/handler.table.update.go
@@ -1,8 +1,0 @@
-package handler
-
-import "github.com/gofiber/fiber/v2"
-
-// TableUpdate handler responsável por tratar os dados e chamar o repo, para realizar updates no registro da mesa.
-func (s DbStruct) TableUpdate(c *fiber.Ctx) error {
-	return nil
-}
