+++ api.event/controller/handler/route.go
@@ -20,10 +20,19 @@ func AllRoutes(s DbStruct, app *fiber.App) {
 	// application/json
 	// application/x-www-form-urlencoded
 	// multipart/form-data
+
+	//Routes Table
 	table.Post("", s.TablePost)
 	table.Get(":offset/:limit", s.TableGetAll)
 	table.Get(":uuid", s.TableGetOne)
 	table.Put("", s.TableUpdate)
 	table.Delete(":uuid", s.TableDelete)
+
+	//Routes Event
+	table.Post("", s.EventPost)
+	table.Get(":offset/:limit", s.EventGetAll)
+	table.Get(":uuid", s.EventGetOne)
+	table.Put("", s.EventUpdate)
+	table.Delete(":uuid", s.EventDelete)
 
 }
